requirements.txt
appnope==0.1.0
backcall==0.1.0
decorator==4.2.1
Django==1.11.12
django-crispy-forms==1.7.2
django-extensions==2.0.6
django-jet==1.0.7
django-rest-framework==0.1.0
djangorestframework==3.8.2
feedparser==5.2.1
google-api-python-client==1.4.1
httplib2==0.11.3
ipython==6.3.1
ipython-genutils==0.2.0
jedi==0.11.1
oauth2client==4.1.2
parso==0.1.1
pexpect==4.4.0
pickleshare==0.7.4
prompt-toolkit==1.0.15
ptyprocess==0.5.2
pyasn1==0.4.2
pyasn1-modules==0.2.1
Pygments==2.2.0
pytz==2018.4
rsa==3.4.2
simplegeneric==0.8.1
six==1.11.0
traitlets==4.3.2
uritemplate==3.0.0
wcwidth==0.1.7
